Translate view with ivec3 argument

+++ src/step_world.cpp
@@ -11,9 +11,9 @@ namespace te {
 
 static void stepView(const glm::vec3& playerTranslation,
 					 glm::mat4& view) {
-	view = glm::translate(glm::vec3(-playerTranslation.x,
-									-playerTranslation.y,
-									0));
+	view = glm::translate(glm::ivec3(-playerTranslation.x,
+									 -playerTranslation.y,
+									 0));
 }
 
 void stepWorld(worldstate_t& state, float dt, const Uint8 *keyboardState, const tilesetstate_t& tilesetState, const levelstate_t& levelState) {
